refactor(`Changes`): rename `ChangeReport` to `Changes`
